fix: jsandoval-2025365 ha arreglado un error con el id (se creaban mas de uno en la base de datos)

+++ src/main/java/com/justinsandoval/entity/Cliente.java
@@ -2,6 +2,8 @@ package com.justinsandoval.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
@@ -19,13 +21,13 @@ public class Cliente {
     @Column
     private String direccion;
     @Column
-    private int estado;
+    private Long estado;
 
     public Cliente(){
 
     }
 
-    public Cliente(String DPICliente, String nombreCliente, String apellidoCliente, String direccion, int estado) {
+    public Cliente(String DPICliente, String nombreCliente, String apellidoCliente, String direccion, Long estado) {
         this.DPICliente = DPICliente;
         this.nombreCliente = nombreCliente;
         this.apellidoCliente = apellidoCliente;
@@ -65,11 +67,11 @@ public class Cliente {
         this.direccion = direccion;
     }
 
-    public int getEstado() {
+    public Long getEstado() {
         return estado;
     }
 
-    public void setEstado(int estado) {
+    public void setEstado(Long estado) {
         this.estado = estado;
     }
 }
